Handling HIDAPI finalizing
hid_exit() finalize the HIDAPI library.

Add check for return value of hid_exit() function responsible for free
"all of the static data associated with HIDAPI".

Rename terminate() function for a better description.

Delete unnecessary call of hid_exit(), because if
an exception is thrown, hid_exit() will be called in the destructor.